Exercises S08 and SinkThem with Tests

+++ src/main/java/rocco/S05.java
@@ -93,7 +93,7 @@ public class S05 {
 		for (int i = 0; i < data.length; i++) {
 			sum += data[i];
 		}
-		return sum/(double)data.length;
+		return sum/data.length;
 	}
 
 	/**
